Heat with loop, 32-bit loops

+++ runtime/cilk_abi_cilk_for_lf.c
@@ -3,7 +3,7 @@
 //
 
 #include <stdio.h>
-#include "stdint.h"
+#include <stdint.h>
 #include "cilk2c.h"
 #include "scheduler.h"
 
@@ -11,7 +11,7 @@ typedef void (*__cilk_abi_f64_t)(void *data, int64_t low, int64_t high);
 
 // we cannot inline this function because of local variables
 static void __attribute__ ((noinline))
-__cilkrts_cilk_loop_helper64(void *data, __cilk_abi_f64_t body, uint64_t grainsize) {
+__cilkrts_cilk_loop_helper64(void *data, __cilk_abi_f64_t body, unsigned int grainsize) {
     uint64_t i;
     __cilkrts_inner_loop_frame inner_lf;
     __cilkrts_enter_inner_loop_frame(&inner_lf);
@@ -37,7 +37,7 @@ __cilkrts_cilk_loop_helper64(void *data, __cilk_abi_f64_t body, uint64_t grainsi
     __cilkrts_leave_frame(&inner_lf.sf);
 }
 
-void __cilkrts_cilk_for_64(__cilk_abi_f64_t body, void *data, uint64_t count, int grain) {
+void __cilkrts_cilk_for_64(__cilk_abi_f64_t body, void *data, uint64_t count, unsigned int grain) {
     __cilkrts_loop_frame lf;
     uint64_t n = count;
     uint64_t end = n / grain, rem = n % grain;
@@ -78,7 +78,7 @@ typedef void (*__cilk_abi_f32_t)(void *data, int32_t low, int32_t high);
 
 // we cannot inline this function because of local variables
 static void __attribute__ ((noinline))
-__cilkrts_cilk_loop_helper32(void *data, __cilk_abi_f32_t body, int grainsize) {
+__cilkrts_cilk_loop_helper32(void *data, __cilk_abi_f32_t body, unsigned int grainsize) {
     uint64_t i;
     __cilkrts_inner_loop_frame inner_lf;
     __cilkrts_enter_inner_loop_frame(&inner_lf);
@@ -104,7 +104,7 @@ __cilkrts_cilk_loop_helper32(void *data, __cilk_abi_f32_t body, int grainsize) {
     __cilkrts_leave_frame(&inner_lf.sf);
 }
 
-void __cilkrts_cilk_for_32(__cilk_abi_f32_t body, void *data, uint32_t count, int grain) {
+void __cilkrts_cilk_for_32(__cilk_abi_f32_t body, void *data, uint32_t count, unsigned int grain) {
     __cilkrts_loop_frame lf;
     uint32_t n = count;
     uint32_t end = n / grain, rem = n % grain;
